+++ Adventure_Time/TextureLoader.h
@@ -7,14 +7,14 @@ class TextureLoader
 {
 public:
 	TextureLoader(const std::string& textureID, int x, int y, int w, int h,
-		int numFrames, double scope = 1.0);
+		int numFrames, float scope = 1.0);
 
 	int getX() const { return mXPos; }
 	int getY() const { return mYPos; }
 	int getWidth() const { return mWidth; }
 	int getHeight() const { return mHeight; }
 	int getNumFrames() const { return mNumFrames; }
-	double getScope() const { return mScope; }
+	float getScope() const { return mScope; }
 	std::string getTextureID() const { return mTextureID; }
 
 private:
@@ -24,7 +24,7 @@ private:
 	int mHeight;
 	int mNumFrames;
 	std::string mTextureID;
-	double mScope;
+	float mScope;
 };
 
 #endif TextureLoader_H_
