+++ Adventure_Time/TextureManager.h
@@ -31,8 +31,8 @@ public:
 	//a: angle to rotate
 	//f: flag to flip
 	void drawSpritePic(const std::string& id, int x, int y,
-		int w, int h, SDL_Renderer* pRenderer = nullptr, int indexFrame = 0, double scope = 1.0,
-		SDL_RendererFlip f = SDL_FLIP_NONE, double a = 0.0, SDL_Point* c = NULL);
+		int w, int h, SDL_Renderer* pRenderer = nullptr, int indexFrame = 0, float scope = 1.0,
+		SDL_RendererFlip f = SDL_FLIP_NONE, float a = 0.0, SDL_Point* c = NULL);
 private:
 	static TextureManager* s_pInstance;
